+++ app/routes/perfil.js
@@ -2,6 +2,7 @@ import { action } from '@ember/object';
 import { inject as service } from '@ember/service';
 import Route from '@ember/routing/route';
 import { hash } from 'rsvp';
+import { typeOf } from '@ember/utils';
 
 const resolver = {
   hospitales: 'institution',
@@ -47,10 +48,12 @@ export default class PerfilRoute extends Route {
         config: {},
         profile: profile,
         compras: this.spreadsheets.fetch('compras').then((compras) => compras.filterBy('id', profile.id).map((compra) => {
-          compra.Monto = parseFloat(compra.Monto.replace('Q','').replace(' ', '').replace(/,/g, ''));
+          if (typeOf(compra.Monto) === 'string') {
+            compra.Monto = parseFloat(compra.Monto.replace('Q','').replace(' ', '').replace(/,/g, ''));
+          }
+
           return compra;
         })),
-        compras2: this.spreadsheets.fetch('compras'),
         consultas: this.spreadsheets.fetch('consultas')
         // profiles: this.store.query('profile', {
         //   institution: profile.id